% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condor_dir.R
\name{condor_dir}
\alias{condor_dir}
\title{Condor Directories}
\usage{
condor_dir(top.dir = "condor", session = NULL)
}
\arguments{
\item{top.dir}{top directory on submitter machine that contains Condor run
directories.}

\item{session}{optional object of class \code{ssh_connect}.}
}
\value{
\code{character} vector of directory names.
}
\description{
List directories on Condor submitter.
}
\details{
The default value of \code{session = NULL} looks for a \code{session} object
in the user workspace. This allows the user to run Condor functions without
explicitly specifying the \code{session}.
}
\examples{
\dontrun{
library(ssh)
session <- ssh_connect("NOUOFPCALC02")

condor_submit()
condor_q()
condor_dir()
condor_download()  # after job has finished
}

}
\seealso{
\code{\link{condor_submit}}, \code{\link{condor_q}}, \code{condor_dir}, and
\code{\link{condor_download}} provide the main Condor interface.

\code{\link{condor-package}} gives an overview of the package.
}
