% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condor_download.R
\name{condor_download}
\alias{condor_download}
\title{Condor Download}
\usage{
condor_download(session, local.dir = ".", remote.dir = "condor_run",
  pattern = "condor_mfcl|End.tar.gz", overwrite.local = FALSE,
  remove.remote = TRUE)
}
\arguments{
\item{session}{created by \code{ssh_connect}.}

\item{local.dir}{local directory to download to.}

\item{remote.dir}{remote directory containing Condor job results.}

\item{pattern}{pattern identifying which result files to download.}

\item{overwrite.local}{whether to overwrite local files if they already
exist.}

\item{remove.remote}{whether to remove the remote directory after downloading
the result files.}
}
\description{
Download results from a Condor job.
}
\examples{
\dontrun{
library(ssh)
session <- ssh_connect("NOUOFPCALC02")

condor_submit(session)
condor_q(session)
condor_download(session)  # after job has finished
}

}
\seealso{
\code{\link{condor_submit}} submits a Condor job.

\code{\link{condor-package}} gives an overview of the package.
}
