% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condor_q.R
\name{condor_q}
\alias{condor_q}
\title{Condor Queue}
\usage{
condor_q(session)
}
\arguments{
\item{session}{created by \code{ssh_connect}.}
}
\description{
List the Condor job queue.
}
\examples{
\dontrun{
library(ssh)
session <- ssh_connect("NOUOFPCALC02")

condor_submit(session)
condor_q(session)
condor_download(session)  # after job has finished
}

}
\seealso{
\code{\link{condor_submit}} submits a Condor job.

\code{\link{condor_download}} downloads results from a Condor job.

\code{\link{condor-package}} gives an overview of the package.
}
