% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condor_rm.R
\name{condor_rm}
\alias{condor_rm}
\title{Condor remove}
\usage{
condor_rm(job.id = "", all = FALSE, session = NULL)
}
\arguments{
\item{job.id}{the id number of the job would like to remove from the condor.}

\item{all}{whether to remove all jobs from users in the condor.}

\item{session}{optional object of class \code{ssh_connect}.}
}
\value{
No return value, called for side effects.
}
\description{
Remove a Condor job.
}
\details{
Use this function to remove one or all the jobs on the condor. 
If wish to remove one job, please specific job.id. 
If wish to remove all jobs, need to set all=TRUE. 
If both job.id and all=FALSE. The function will stop and return an error message.
}
\examples{
\dontrun{

# General workflow
session <- ssh_connect("servername")

condor_submit()
condor_q()
condor_rm() # if want to remove a job from condor
condor_dir()
condor_download()  # after job has finished
condor_rmdir()

# Alternatively, list number of jobs being run by each user
condor_q(all=TRUE, count=TRUE)
}

}
\author{
Nan Yao.

 @seealso
\code{\link{condor_submit}}, \code{\link{condor_q}},
\code{\link{condor_rm}}, \code{\link{condor_dir}}, 
\code{condor_download}, and
\code{\link{condor_rmdir}} provide the main Condor interface.

\code{\link{condor-package}} gives an overview of the package.
}
