% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condor_rmdir.R
\name{condor_rmdir}
\alias{condor_rmdir}
\title{Condor Remove Directory}
\usage{
condor_rmdir(run.dir = NULL, top.dir = "condor", session = NULL)
}
\arguments{
\item{run.dir}{name of a Condor run directory inside \code{top.dir}.}

\item{top.dir}{top directory on submitter machine that contains Condor run
directories.}

\item{session}{optional object of class \code{ssh_connect}.}
}
\value{
No return value, called for side effects.
}
\description{
Remove a directory on the submitter machine.
}
\details{
The default value of \code{run.dir = NULL} looks for a Condor run directory
with the same name as the current working directory:
\emph{top.dir}\code{/}\emph{working.dir}. For example, if the current working
directory is \code{"c:/yft/run01"} then the default \code{run.dir} becomes
\code{"condor/run01"}.

The default value of \code{session = NULL} looks for a \code{session} object
in the user workspace. This allows the user to run Condor functions without
explicitly specifying the \code{session}.
}
\examples{
\dontrun{

# General workflow
session <- ssh_connect("servername")

condor_submit()
condor_q()
condor_dir()
condor_download()  # after job has finished
condor_rmdir()

# Alternatively, remove a specific directory
condor_rmdir("01_this_model")           # remove ~/condor/01_this_model
condor_rmdir("test_runs", top.dir=".")  # remove ~/test_runs
}

}
\seealso{
\code{\link{condor_submit}}, \code{\link{condor_q}},
\code{\link{condor_dir}}, \code{condor_download}, and
\code{\link{condor_rmdir}} provide the main Condor interface.

\code{\link{condor-package}} gives an overview of the package.
}
