% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condor_submit.R
\name{condor_submit}
\alias{condor_submit}
\title{Condor Submit}
\usage{
condor_submit(session, local.dir = ".", remote.dir = "condor_run")
}
\arguments{
\item{session}{created by \code{ssh_connect}.}

\item{local.dir}{local directory containing \code{condor.sub} and any other
files necessary to run the job.}

\item{remote.dir}{remote directory containing Condor job results.}
}
\description{
Submit a Condor job.
}
\examples{
\dontrun{
library(ssh)
session <- ssh_connect("NOUOFPCALC02")

condor_submit(session)
condor_q(session)
condor_download(session)  # after job has finished
}

}
\seealso{
\code{\link{condor_download}} downloads results from a Condor job.

\code{\link{condor-package}} gives an overview of the package.
}
