% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dos2unix.R, R/unix2dos.R
\name{dos2unix}
\alias{dos2unix}
\alias{unix2dos}
\title{Convert Line Endings}
\usage{
dos2unix(file)

unix2dos(file)
}
\arguments{
\item{file}{a filename.}
}
\value{
No return value, called for side effects.
}
\description{
Convert line endings in a text file between Dos (CRLF) and Unix (LF) format.
}
\examples{
\dontrun{
file <- "test.txt"
write("123", file)

dos2unix(file)
file.size(file)

unix2dos(file)
file.size(file)

file.remove(file)
}

}
\seealso{
\code{\link{condor-package}} gives an overview of the package.
}
\author{
Arni Magnusson.
}
