% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condor-internal.R, R/condor_log.R
\name{condor-internal}
\alias{condor-internal}
\alias{print.condor_log}
\title{Various Condor Helper Functions}
\usage{
\method{print}{condor_log}(x, ...)
}
\value{
No return value, called for side effects.
}
\description{
These functions are called by user-level functions. The functionality is
documented in the user-level functions.
}
\seealso{
\code{\link{condor-package}} gives an overview of the package.
}
\keyword{internal}
