% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condor-package.R
\docType{package}
\name{condor-package}
\alias{condor-package}
\alias{condor}
\title{Interact with Condor from R}
\description{
Tools that support submitting, listing, and downloading Condor jobs from R.
}
\details{
\emph{Condor jobs:}
\tabular{ll}{
  \code{\link{condor_download}} \tab download\cr
  \code{\link{condor_q}}        \tab list queue\cr
  \code{\link{condor_dir}}      \tab list directories\cr
  \code{\link{condor_submit}}   \tab submit
}
\emph{Utilities:}
\tabular{ll}{
  \code{\link{ssh_exec_stdout}} \tab execute command
}
}
\references{
\url{https://github.com/PacificCommunity/ofp-sam-condor}
}
\seealso{
\pkg{condor} uses the \pkg{ssh} package to connect to the Condor submitter
machine.
}
\author{
Arni Magnusson with contributions by Jemery Day.
}
