% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condor_dir.R
\name{condor_dir}
\alias{condor_dir}
\title{Condor Directories}
\usage{
condor_dir(topdir = "condor", session = NULL)
}
\arguments{
\item{topdir}{top directory on submitter machine that contains Condor run
directories.}

\item{session}{optional object of class \code{ssh_connect}.}
}
\description{
List directories on Condor submitter machine.
}
\details{
The default value of \code{session = NULL} looks for a \code{session} object
in the user workspace. This allows the user to run Condor functions without
explicitly specifying the \code{session}.
}
\examples{
\dontrun{
library(ssh)
session <- ssh_connect("NOUOFPCALC02")

condor_dir()
}

}
\seealso{
\code{\link{condor_submit}} submits a Condor job.

\code{\link{condor_q}} lists the Condor job queue.

\code{\link{condor_download}} downloads results from a Condor job.

\code{\link{condor-package}} gives an overview of the package.
}
