% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condor_download.R
\name{condor_download}
\alias{condor_download}
\title{Condor Download}
\usage{
condor_download(run.dir = NULL, top.dir = "condor", local.dir = ".",
  subdir = FALSE, pattern = "End.tar.gz|condor.*(err|log|out)$",
  overwrite = FALSE, remove = TRUE, untar.end = TRUE, session = NULL)
}
\arguments{
\item{run.dir}{name of a Condor run directory inside \code{top.dir}.}

\item{top.dir}{top directory on submitter machine that contains Condor run
directories.}

\item{local.dir}{local directory to download to, possibly combined with a
\code{subdir}.}

\item{subdir}{subdirectory to append to \code{local.dir}.}

\item{pattern}{regular expression identifying which result files to download.}

\item{overwrite}{whether to overwrite local files if they already exist.}

\item{remove}{whether to remove remote directory after downloading result
files.}

\item{untar.end}{whether to extract \code{End.tar.gz} into
\emph{local.dir}\code{/End}. (Ignored if a file named
\file{End.tar.gz} was not downloaded.)}

\item{session}{optional object of class \code{ssh_connect}.}
}
\description{
Download results from a Condor job.
}
\details{
The default value of \code{run.dir = NULL} looks for Condor job results in
\emph{top.dir}\code{/}\emph{local.dir}. For example, if
\code{local.dir = "c:/yft/run01"} then the default \code{run.dir} becomes
\code{"condor/run01"}.

Generally, \code{local.dir} is updated by appending \code{subdir}. To
download directly to \code{local.dir} instead, pass \code{subdir = FALSE},
\code{NULL}, or \code{""}. If \code{local.dir} already ends with
\code{subdir}, then \code{subdir} is ignored.

The default value of \code{session = NULL} looks for a \code{session} object
in the user workspace. This allows the user to run Condor functions without
explicitly specifying the \code{session}.
}
\examples{
\dontrun{
library(ssh)
session <- ssh_connect("NOUOFPCALC02")

condor_submit()
condor_q()
condor_dir()
condor_download()  # after job has finished
}

}
\seealso{
\code{\link{condor_submit}}, \code{\link{condor_q}},
\code{\link{condor_dir}}, and \code{condor_download} provide the main Condor
interface.

\code{\link{condor-package}} gives an overview of the package.
}
