% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condor_download.R
\name{condor_download}
\alias{condor_download}
\title{Condor Download}
\usage{
condor_download(local.dir = ".", remote.dir = "condor_run",
  pattern = "condor_mfcl|End.tar.gz", overwrite.local = FALSE,
  remove.remote = TRUE, session = NULL)
}
\arguments{
\item{local.dir}{local directory to download to.}

\item{remote.dir}{remote directory containing Condor job results.}

\item{pattern}{pattern identifying which result files to download.}

\item{overwrite.local}{whether to overwrite local files if they already
exist.}

\item{remove.remote}{whether to remove the remote directory after downloading
the result files.}

\item{session}{optional object of class \code{ssh_connect}.}
}
\description{
Download results from a Condor job.
}
\details{
When the default value of \code{session = NULL} is used, the function looks
for a \code{session} object in the user workspace. This allows the user to
run Condor functions without explicitly specifying the \code{session}.
}
\examples{
\dontrun{
library(ssh)
session <- ssh_connect("NOUOFPCALC02")

condor_submit()
condor_q()
condor_download()  # after job has finished
}

}
\seealso{
\code{\link{condor_submit}} submits a Condor job.

\code{\link{condor_q}} lists the Condor job queue.

\code{\link{condor-package}} gives an overview of the package.
}
