% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condor_q.R
\name{condor_q}
\alias{condor_q}
\title{Condor Queue}
\usage{
condor_q(session = NULL)
}
\arguments{
\item{session}{optional object of class \code{ssh_connect}.}
}
\description{
List the Condor job queue.
}
\details{
The default value of \code{session = NULL} looks for a \code{session} object
in the user workspace. This allows the user to run Condor functions without
explicitly specifying the \code{session}.
}
\examples{
\dontrun{
library(ssh)
session <- ssh_connect("NOUOFPCALC02")

condor_submit()
condor_q()
condor_download()  # after job has finished
}

}
\seealso{
\code{\link{condor_submit}} submits a Condor job.

\code{\link{condor_download}} downloads results from a Condor job.

\code{\link{condor-package}} gives an overview of the package.
}
