% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condor_submit.R
\name{condor_submit}
\alias{condor_submit}
\title{Condor Submit}
\usage{
condor_submit(remote.dir = basename(local.dir), local.dir = getwd(),
  exclude = "condor_mfcl|tar.gz|End", session = NULL)
}
\arguments{
\item{remote.dir}{remote directory where Condor job should run.}

\item{local.dir}{local directory containing a Condor \code{*.sub} file and
any other files necessary to run the job.}

\item{exclude}{pattern identifying files that should not be submitted to
Condor.}

\item{session}{optional object of class \code{ssh_connect}.}
}
\description{
Submit a Condor job.
}
\details{
The default value of \code{remote.dir} uses the same directory name as
\code{local.dir}.

The default value of \code{session = NULL} looks for a \code{session} object
in the user workspace. This allows the user to run Condor functions without
explicitly specifying the \code{session}.
}
\examples{
\dontrun{
library(ssh)
session <- ssh_connect("NOUOFPCALC02")

condor_submit()
condor_q()
condor_download()  # after job has finished
}

}
\seealso{
\code{\link{condor_q}} lists the Condor job queue.

\code{\link{condor_download}} downloads results from a Condor job.

\code{\link{condor-package}} gives an overview of the package.
}
